requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
black==21.9b0
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudscraper==1.2.58
emoji==1.6.1
idna==3.3
lxml==4.6.3
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
regex==2021.10.23
requests==2.26.0
requests-toolbelt==0.9.1
soupsieve==2.2.1
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
